Move service-like functions from conductor.tasks to conductor.service
