+++ concent_api/conductor/service.py
@@ -1,11 +1,16 @@
 import logging
+from typing import List
+from typing import Optional
 
 from django.db.models import Q
 
 from common.constants import ErrorCode
+from common.helpers import parse_timestamp_to_utc_datetime
 from common.logging import log
 from common.logging import LoggingLevel
 from conductor.exceptions import VerificationRequestAlreadyInitiatedError
+from conductor.models import BlenderSubtaskDefinition
+from conductor.models import Frame
 from conductor.models import ResultTransferRequest
 from conductor.models import UploadReport
 from conductor.models import VerificationRequest
@@ -54,3 +59,50 @@ def update_upload_report(file_path: str, result_transfer_request: ResultTransfer
         result_transfer_request.save()
 
         result_upload_finished.delay(result_transfer_request.subtask_id)
+
+
+def store_verification_request_and_blender_subtask_definition(
+    subtask_id: str,
+    source_package_path: str,
+    result_package_path: str,
+    output_format: str,
+    scene_file: str,
+    verification_deadline: int,
+    blender_crop_script: Optional[str],
+) -> tuple:
+    verification_request = VerificationRequest(
+        subtask_id=subtask_id,
+        source_package_path=source_package_path,
+        result_package_path=result_package_path,
+        verification_deadline=parse_timestamp_to_utc_datetime(verification_deadline),
+    )
+    verification_request.full_clean()
+    verification_request.save()
+
+    blender_subtask_definition = BlenderSubtaskDefinition(
+        verification_request=verification_request,
+        output_format=BlenderSubtaskDefinition.OutputFormat[output_format].name,
+        scene_file=scene_file,
+        blender_crop_script=blender_crop_script,
+    )
+    blender_subtask_definition.full_clean()
+    blender_subtask_definition.save()
+
+    return (verification_request, blender_subtask_definition)
+
+
+def store_frames(
+    blender_subtask_definition: BlenderSubtaskDefinition,
+    frame_list: List[int],
+) -> None:
+    for frame in frame_list:
+        store_frame = Frame(
+            blender_subtask_definition=blender_subtask_definition,
+            number=frame,
+        )
+        store_frame.full_clean()
+        store_frame.save()
+
+
+def filter_frames_by_blender_subtask_definition(blender_subtask_definition: BlenderSubtaskDefinition) -> list:
+    return list(Frame.objects.filter(blender_subtask_definition=blender_subtask_definition).values_list('number', flat=True))
