+++ concent_api/conductor/tasks.py
@@ -10,15 +10,16 @@ from common.constants import ErrorCode
 from common.decorators import log_task_errors
 from common.decorators import provides_concent_feature
 from common.helpers import parse_datetime_to_timestamp
-from common.helpers import parse_timestamp_to_utc_datetime
 from common.logging import log
 from common.logging import LoggingLevel
 from conductor.exceptions import VerificationRequestAlreadyAcknowledgedError
 from conductor.models import BlenderSubtaskDefinition
-from conductor.models import Frame
 from conductor.models import ResultTransferRequest
 from conductor.models import UploadReport
 from conductor.models import VerificationRequest
+from conductor.service import filter_frames_by_blender_subtask_definition
+from conductor.service import store_frames
+from conductor.service import store_verification_request_and_blender_subtask_definition
 from conductor.service import update_upload_report
 from core import tasks
 from core.validation import validate_frames
@@ -180,53 +181,6 @@ def upload_acknowledged(
         f'Result_package_hash: {result_package_hash}',
         subtask_id=subtask_id
     )
-
-
-def store_verification_request_and_blender_subtask_definition(
-    subtask_id: str,
-    source_package_path: str,
-    result_package_path: str,
-    output_format: str,
-    scene_file: str,
-    verification_deadline: int,
-    blender_crop_script: Optional[str],
-) -> tuple:
-    verification_request = VerificationRequest(
-        subtask_id=subtask_id,
-        source_package_path=source_package_path,
-        result_package_path=result_package_path,
-        verification_deadline=parse_timestamp_to_utc_datetime(verification_deadline),
-    )
-    verification_request.full_clean()
-    verification_request.save()
-
-    blender_subtask_definition = BlenderSubtaskDefinition(
-        verification_request=verification_request,
-        output_format=BlenderSubtaskDefinition.OutputFormat[output_format].name,
-        scene_file=scene_file,
-        blender_crop_script=blender_crop_script,
-    )
-    blender_subtask_definition.full_clean()
-    blender_subtask_definition.save()
-
-    return (verification_request, blender_subtask_definition)
-
-
-def store_frames(
-    blender_subtask_definition: BlenderSubtaskDefinition,
-    frame_list: List[int],
-) -> None:
-    for frame in frame_list:
-        store_frame = Frame(
-            blender_subtask_definition=blender_subtask_definition,
-            number=frame,
-        )
-        store_frame.full_clean()
-        store_frame.save()
-
-
-def filter_frames_by_blender_subtask_definition(blender_subtask_definition: BlenderSubtaskDefinition) -> list:
-    return list(Frame.objects.filter(blender_subtask_definition=blender_subtask_definition).values_list('number', flat=True))
 
 
 @shared_task
